escola2/requirements.txt
asgiref==3.8.1
autocommand==2.2.2
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
docbr==0.1.3
drf-yasg==1.21.7
Faker==26.0.0
inflect==7.4.0
inflection==0.5.1
jaraco.context==6.0.1
jaraco.functools==4.0.2
jaraco.itertools==6.4.1
jaraco.text==4.0.0
Markdown==3.6
more-itertools==10.5.0
numpy==2.0.0
packaging==24.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
validate-docbr==1.10.0
